add tets

+++ tests/phound_test/test.sh
@@ -18,6 +18,19 @@ if ! type sqlite3 >/dev/null; then
     exit 1
 fi
 
+EXPECTED_PATH=expected/all_output.expected
+if [ ! -d expected  ]; then
+    echo "Error: must run this script from tests/override_return_types folder" 1>&2
+    exit 1
+fi
+echo "Generating test cases"
+for path in $(echo expected/*.php.expected | LC_ALL=C sort); do cat "$path"; done > $EXPECTED_PATH
+EXIT_CODE=$?
+if [[ $EXIT_CODE != 0 ]]; then
+    echo "Failed to concatenate test cases" 1>&2
+    exit 1
+fi
+
 echo "Running phan in '$PWD' ..."
 
 rm -rf ~/phound.db
@@ -28,62 +41,20 @@ if ! ../../phan --force-polyfill-parser --memory-limit 1G --analyze-twice ; then
     exit 1
 fi
 
-# Regarding `order by cast(substr(callsite, instr(callsite, ":") + 1) as integer)` -
-# This orders by the callsite line number. It avoids weirdness where, for example,
-# 'foo.php:10' might otherwise appear ahead of 'foo.php:9' when treated as a string.
-ACTUAL=$(sqlite3 ~/phound.db 'select * from callsites order by cast(substr(callsite, instr(callsite, ":") + 1) as integer), element, type')
-EXPECTED=$(cat <<-EOF
-\A::foo|const|src/001_phound_callsites.php:10
-\A::foo|prop|src/001_phound_callsites.php:11
-\A::fooz|prop|src/001_phound_callsites.php:12
-\A::foo|method|src/001_phound_callsites.php:13
-\A::bar|method|src/001_phound_callsites.php:14
-\B::__construct|method|src/001_phound_callsites.php:19
-\C::__construct|method|src/001_phound_callsites.php:21
-\A::getBOrC|method|src/001_phound_callsites.php:34
-\B::foo|method|src/001_phound_callsites.php:35
-\C::foo|method|src/001_phound_callsites.php:35
-\B::bar|prop|src/001_phound_callsites.php:36
-\C::bar|prop|src/001_phound_callsites.php:36
-\A::getBOrCClassName|method|src/001_phound_callsites.php:38
-\B::zoo|method|src/001_phound_callsites.php:39
-\C::zoo|method|src/001_phound_callsites.php:39
-\B::baz|prop|src/001_phound_callsites.php:40
-\C::baz|prop|src/001_phound_callsites.php:40
-\B::BOO|const|src/001_phound_callsites.php:41
-\C::BOO|const|src/001_phound_callsites.php:41
-\TestConstructor::__construct|method|src/001_phound_callsites.php:71
-\A::__construct|method|src/001_phound_callsites.php:75
-\A::foo|method|src/001_phound_callsites.php:76
-\A::bar|method|src/001_phound_callsites.php:77
-\A::foo|method|src/001_phound_callsites.php:79
-\A::getBOrC|method|src/001_phound_callsites.php:80
-\B::foo|method|src/001_phound_callsites.php:80
-\C::foo|method|src/001_phound_callsites.php:80
-\A::bar|method|src/001_phound_callsites.php:81
-\A::baz|method|src/001_phound_callsites.php:82
-\A::__construct|method|src/001_phound_callsites.php:86
-\A::bar|method|src/001_phound_callsites.php:86
-\A::getBOrCClassName|method|src/001_phound_callsites.php:88
-\B::__construct|method|src/001_phound_callsites.php:89
-\B::foo|method|src/001_phound_callsites.php:89
-\C::__construct|method|src/001_phound_callsites.php:89
-\C::foo|method|src/001_phound_callsites.php:89
-\A::foo|method|src/001_phound_callsites.php:91
-\Closure::fromCallable|method|src/001_phound_callsites.php:91
-\A::getBOrC|method|src/001_phound_callsites.php:94
-\B::foo|method|src/001_phound_callsites.php:94
-\C::foo|method|src/001_phound_callsites.php:94
-\Closure::fromCallable|method|src/001_phound_callsites.php:94
-\A::getBOrC|method|src/001_phound_callsites.php:97
-\B::methodOnlyDefinedInB|method|src/001_phound_callsites.php:99
-\B::prop_only_public_in_b|prop|src/001_phound_callsites.php:100
-EOF
-)
+# Regarding the ORDER BY clause:
+# 1) `substr(callsite, 0, instr(callsite, ":"))`
+#   This transforms a callsite like `001_my_test_file.php:10`` to `001_my_test_file.php`
+# 2) `cast(substr(callsite, instr(callsite, ":") + 1) as integer)`
+#   This transforms a callsite like 001_my_test_file.php:10 to `10`.
+#   That is, it transforms it to the callsite line number as an integer. It avoids weirdness where,
+#   for example, 'foo.php:10' might otherwise appear ahead of 'foo.php:9' when treated as a string.
+#
+# Together, these order by clauses ensure the output is ordered by file and line number.
+ACTUAL=$(sqlite3 ~/phound.db 'SELECT * FROM callsites ORDER BY substr(callsite, 0, instr(callsite, ":")), cast(substr(callsite, instr(callsite, ":") + 1) as integer), element, type')
 
 # diff returns a non-zero exit code if files differ or are missing
 # This outputs the difference between actual and expected output.
-echo
+echo "$ACTUAL"
 echo "Comparing the output:"
 
 if type colordiff >/dev/null; then
@@ -92,7 +63,7 @@ else
     DIFF="diff"
 fi
 
-$DIFF <(echo "$EXPECTED") <(echo "$ACTUAL")
+$DIFF $EXPECTED_PATH <(echo "$ACTUAL")
 EXIT_CODE=$?
 if [ "$EXIT_CODE" == 0 ]; then
     echo "The sqlite3 DB content matches what was expected"
